requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=0.1.0
langchain-community>=0.2.0
neo4j>=5.20.0
requests
numpy
python-dotenv>=1.0.0
openai>=1.0.0
tqdm
datasets
